refactor: move remover logic to own file
