--- src/context.ts
@@ -1,11 +1,0 @@
-export type ListenerRemover = () => void;
-
-export interface Context {
-  listenerRemovers: {
-    [key: string]: ListenerRemover;
-  };
-}
-
-const context: Context = { listenerRemovers: {} };
-
-export default context;
