+++ src/utils/events/eventListeners/index.ts
@@ -1,2 +1,3 @@
 export * from './eventListeners';
 export * from './eventListenerGenerator';
+export * from './listenerRemover';
