+++ src/utils/motions.ts
@@ -1,15 +1,9 @@
-import { AnyColonyClient, ClientType, Extension } from '@colony/colony-js';
+import { AnyColonyClient, Extension } from '@colony/colony-js';
 import { TransactionDescription } from 'ethers/lib/utils';
-import ctx, { ListenerRemover } from '~context';
 
 import { mutate } from '../amplifyClient';
-import {
-  ContractEvent,
-  ContractEventsSignatures,
-  motionNameMapping,
-} from '../types';
+import { ContractEvent, motionNameMapping } from '../types';
 import { getDomainDatabaseId } from './domains';
-import { generateListenerRemoverKey } from './events';
 import { verbose } from './logger';
 import { getColonyTokenAddress } from './tokens';
 
@@ -60,36 +54,5 @@ export const writeMintTokensMotionToDB = async (
       amount,
       blockNumber,
     },
-  });
-};
-
-const getMotionListenerRemovers = (
-  colonyAddress: string,
-): ListenerRemover[] => {
-  const listenerRemovers: ListenerRemover[] = [];
-
-  const motionSignatures = Object.values(ContractEventsSignatures).filter(
-    (sig) => sig.includes('Motion'),
-  );
-
-  motionSignatures.forEach((signature) => {
-    const key = generateListenerRemoverKey(
-      colonyAddress,
-      ClientType.VotingReputationClient,
-      signature,
-    );
-    const motionListenerRemover = ctx.listenerRemovers[key];
-    if (motionListenerRemover) {
-      listenerRemovers.push(motionListenerRemover);
-    }
-  });
-
-  return listenerRemovers;
-};
-
-export const removeMotionListeners = (colonyAddress: string): void => {
-  const listenerRemovers = getMotionListenerRemovers(colonyAddress);
-  listenerRemovers.forEach((listenerRemover) => {
-    listenerRemover();
   });
 };
